use right type for len.
2005-08-23  Not Zed  <NotZed@Ximian.com>

        * test-suite.c (plain_save_receiver): use right type for len.

        * gtkhtml.c (drag_data_get): use right type for g_convert.

2005-01-11  Not Zed  <NotZed@Ximian.com>

        * Makefile.am: put the version in pkgconfig_DATA and remove the
        install-data-hook.  This was breaking on amd64 systems with a
        different libdir.

svn path=/trunk/; revision=7976

+++ src/gtkhtml.c
@@ -2510,7 +2510,7 @@ drag_data_get (GtkWidget *widget, GdkDragContext *context, GtkSelectionData *sel
 					 */
 					char *ucs2;
 					char *utf8;
-					int written_len;
+					gsize written_len;
 
 					if (HTML_IS_TEXT (obj)) {
 						Link *link = html_text_get_link_at_offset (HTML_TEXT (obj), offset);
